+++ osidb/models.py
@@ -1316,7 +1316,7 @@ class Flaw(
         # TODO we have no tracker resolution for now
         return False
 
-    def bzsync(self, *args, bz_api_key, **kwargs):
+    def bzsync(self, *args, force_synchronous_sync=False, **kwargs):
         """
         Bugzilla sync of the Flaw instance
         """
@@ -1331,17 +1331,17 @@ class Flaw(
             return
 
         # switch of sync/async processing
-        if SYNC_FLAWS_TO_BZ_ASYNCHRONOUSLY:
+        if SYNC_FLAWS_TO_BZ_ASYNCHRONOUSLY and not force_synchronous_sync:
             # Process the bzsync asynchronously
             BZSyncManager.check_for_reschedules()
-            BZSyncManager.schedule(str(self.uuid), *args, **kwargs)
+            BZSyncManager.schedule(str(self.uuid))
         else:
             # imports here to prevent cycles
             from apps.bbsync.save import FlawBugzillaSaver
             from collectors.bzimport.collectors import FlawCollector
 
             # sync to Bugzilla
-            bs = FlawBugzillaSaver(self, bz_api_key)
+            bs = FlawBugzillaSaver(self, kwargs.get("bz_api_key"))
             flaw_instance = bs.save()
             kwargs[
                 "auto_timestamps"
@@ -2130,13 +2130,13 @@ class Affect(
             return None
         return ps_module.ps_product.name
 
-    def bzsync(self, *args, bz_api_key, **kwargs):
+    def bzsync(self, *args, **kwargs):
         """
         Bugzilla sync of the Affect instance
         """
         self.save()
         # Affect needs to be synced through flaw
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.flaw.save(*args, **kwargs)
 
 
 class FlawCVSSManager(ACLMixinManager, TrackingMixinManager):
@@ -2259,13 +2259,13 @@ class FlawCVSS(CVSS):
             GinIndex(fields=["acl_read"]),
         ]
 
-    def bzsync(self, *args, bz_api_key, **kwargs):
+    def bzsync(self, *args, **kwargs):
         """
         Bugzilla sync of the FlawCVSS instance
         """
         self.save()
         # FlawCVSS needs to be synced through flaw
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.flaw.save(*args, **kwargs)
 
 
 class AffectCVSS(CVSS):
@@ -2285,13 +2285,13 @@ class AffectCVSS(CVSS):
             GinIndex(fields=["acl_read"]),
         ]
 
-    def bzsync(self, *args, bz_api_key, **kwargs):
+    def bzsync(self, *args, **kwargs):
         """
         Bugzilla sync of the AffectCVSS instance
         """
         self.save()
         # AffectCVSS needs to be synced through affect
-        self.affect.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.affect.save(*args, **kwargs)
 
 
 class TrackerManager(ACLMixinManager, TrackingMixinManager):
@@ -2862,7 +2862,11 @@ class FlawComment(
         # If external_system_id is blank, BugzillaSaver posts the new comment
         # and FlawCollector loads the new comment and updates this FlawComment
         # instance to match bugzilla.
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        # NOTE: Keep using user BZ API key for Flaw comments as we need to
+        #       preserve the information about author
+        self.flaw.save(
+            *args, bz_api_key=bz_api_key, force_synchronous_sync=True, **kwargs
+        )
 
 
 class FlawAcknowledgmentManager(ACLMixinManager, TrackingMixinManager):
@@ -2948,7 +2952,7 @@ class FlawAcknowledgment(AlertMixin, ACLMixin, BugzillaSyncMixin, TrackingMixin)
         """
         self.save()
         # FlawAcknowledgment needs to be synced through flaw
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.flaw.save(*args, **kwargs)
 
 
 class FlawReferenceManager(ACLMixinManager, TrackingMixinManager):
@@ -3068,7 +3072,7 @@ class FlawReference(AlertMixin, ACLMixin, BugzillaSyncMixin, TrackingMixin):
         """
         self.save()
         # FlawReference needs to be synced through flaw
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.flaw.save(*args, **kwargs)
 
 
 class VersionStatus(models.TextChoices):
@@ -3206,7 +3210,7 @@ class Package(AlertMixin, ACLMixin, BugzillaSyncMixin, TrackingMixin):
         """
         self.save()
         # needs to be synced through flaw
-        self.flaw.save(*args, bz_api_key=bz_api_key, **kwargs)
+        self.flaw.save(*args, **kwargs)
 
     class Meta:
         """define meta"""